Data.Grid

Description

This module defines Grid datatype. Feel free to define additional helper functions to work with this datatype in this module.

Documentation

newtype Grid a

Grid. Has a method, that returns a ListZipper (LZ) of LZ's

Constructors

Grid

Fields

unGrid :: ListZipper (ListZipper a)

gUp :: Grid a -> Grid a

Returns a grid, that can be achieved from the input grid, by moving the 'current element' up or down, which is achived by shifting the external LZ left or right.

gDown :: Grid a -> Grid a

Returns a grid, that can be achieved from the input grid, by moving the 'current element' up or down, which is achived by shifting the external LZ left or right.

gLeft :: Grid a -> Grid a

Returns a grid, that can be achieved from the input grid, by moving the 'current element' left or right, which is achieved by shifting all of the internal LZs left or right

gRight :: Grid a -> Grid a

Returns a grid, that can be achieved from the input grid, by moving the 'current element' left or right, which is achieved by shifting all of the internal LZs left or right

gWrite :: a -> Grid a -> Grid a

Change the 'current element'

gHorizontal :: Grid a -> ListZipper (Grid a)

Generate a LZ of all horizontal or vertical 'current element' moves

gVertical :: Grid a -> ListZipper (Grid a)

Generate a LZ of all horizontal or vertical 'current element' moves