HW6.T1

Documentation

type BucketsArray stm k v = TArray stm Int (Bucket k v)

Type alias for an array of lists of Key-ValuePair The hash of the key will be the index of this array where the element may be stored

data CHT stm k v

Concurrent Hash Table data type

Constructors

CHT

Fields

chtBuckets :: TVar stm (BucketsArray stm k v)

chtSize :: TVar stm Int

newCHT :: MonadConc m => m (CHT (STM m) k v)

Creates an empty Hash Table

getCHT :: (MonadConc m, Key k) => k -> CHT (STM m) k v -> m (Maybe v)

Gets Value from Key in CHT. If the key isn't found, returns Nothing

putCHT :: (MonadConc m, Key k) => k -> v -> CHT (STM m) k v -> m ()

Puts Key Value pair into CHT

sizeCHT :: MonadConc m => CHT (STM m) k v -> m Int

returns the amount of elements that are put into the CHT

initCapacity :: Int

Initial capacity of CHT

loadFactor :: Double

This is the threshold for when we need to update the capacity of our CHT. chtSize >= capacity * loadFactor then resize