HW6.T3

Documentation

data Config

data type for configuring wanted properties of our virus i.e. how easily it spreads (probability), the incubationPeriod, illnessDuration and immunityDuration

Constructors

Config

Fields

probability :: Double

incubationPeriod :: Int

illnessDuration :: Int

immunityDuration :: Int

data Cell

Cell contains it's state and a generator for creating random numbers

Constructors

Cell

Fields

cellState :: CellState

cellRand :: StdGen

data CellState

A Cell can be Healthy, Infected, Ill or Imune

Constructors

Healthy

Infected Int

Ill Int

Immune Int

type Comonad19Grid = Grid Cell

Type alias for a Grid containing Cells

printGrid :: Int -> Comonad19Grid -> IO ()

Takes in the size (wanted view) of the grid and prints it

simulate :: Int -> Config -> [Comonad19Grid]

Creates an infinite list of grids using the given configuration. Each element of this list represents one infection simulation step. Takes in a seed and a config